# packages/workflow-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowAId API",
    "version": "1.1.0",
    "description": "Typed decision workflows: build, run, observe and evaluate. See docs/design/API.md."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "`fa_live_…` / `fa_test_…` API key"
      },
      "session": { "type": "apiKey", "in": "cookie", "name": "__Host-fa_session" }
    },
...
  },
  "paths": {
...
    "/v1/webhooks": {
      "get": {
        "tags": ["triggers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "workflowId",
            "required": false
          }
        ],
        "x-cli": { "noun": "webhook", "verb": "list" },
        "x-auth": "session_or_api_key",
        "x-scopes": ["webhooks:write"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/webhooks/{id}": {
      "patch": {
        "tags": ["triggers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "secretCredentialId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      { "type": "null" }
                    ]
                  },
                  "callbackUrl": {
                    "anyOf": [{ "type": "string", "format": "uri" }, { "type": "null" }]
                  },
                  "allowedHeaders": {
                    "maxItems": 50,
                    "type": "array",
                    "items": { "type": "string", "maxLength": 100 }
                  },
                  "enabled": { "type": "boolean" },
                  "requireTimestamp": { "type": "boolean" },
                  "idempotencyHeader": {
                    "anyOf": [{ "type": "string", "maxLength": 100 }, { "type": "null" }]
                  },
                  "path": { "type": "string" }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "webhook", "verb": "update", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["webhooks:write"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/webhooks/{id}/rotate-secret": {
      "post": {
        "summary": "Generate a new signing secret (returned once)",
        "tags": ["triggers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "webhook", "verb": "rotate-secret", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["webhooks:write"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": { "type": "string" },
                    "credentialId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["secret", "credentialId"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/schedules": {
      "get": {
        "tags": ["triggers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "workflowId",
            "required": false
          }
        ],
        "x-cli": { "noun": "schedule", "verb": "list" },
        "x-auth": "session_or_api_key",
        "x-scopes": ["schedules:write"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/schedules/{id}": {
      "patch": {
        "tags": ["triggers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" },
                  "overlap": { "type": "string", "enum": ["skip", "allow"] },
                  "catchUp": { "type": "string", "enum": ["skip", "one", "all"] },
                  "maxCatchUp": { "type": "integer", "minimum": 1, "maximum": 100 },
                  "jitterMs": { "type": "integer", "minimum": 0, "maximum": 3600000 },
                  "cron": { "type": "string" }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "schedule", "verb": "update", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["schedules:write"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/schedules/{id}/trigger": {
      "post": {
        "summary": "Fire the schedule now (its input, its environment)",
        "tags": ["triggers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "schedule", "verb": "trigger", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["schedules:write"],
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "run_id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["run_id"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit": {
      "get": {
        "tags": ["audit"],
        "parameters": [
          {
            "schema": { "default": 50, "type": "integer", "minimum": 1, "maximum": 200 },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "string", "maxLength": 500 },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": { "default": "desc", "type": "string", "enum": ["asc", "desc"] },
            "in": "query",
            "name": "order",
            "required": false
          },
          { "schema": { "type": "string" }, "in": "query", "name": "actor", "required": false },
          { "schema": { "type": "string" }, "in": "query", "name": "action", "required": false },
          {
            "schema": { "type": "string" },
            "in": "query",
            "name": "resourceType",
            "required": false
          },
          {
            "schema": { "type": "string" },
            "in": "query",
            "name": "resourceId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "x-cli": { "noun": "audit", "verb": "list" },
        "x-auth": "session_or_api_key",
        "x-scopes": ["audit:read"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/artifacts/{id}/download": {
      "get": {
        "tags": ["artifacts"],
        "parameters": [
          {
            "schema": { "default": false, "type": "boolean" },
            "in": "query",
            "name": "inline",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "x-cli": { "noun": "artifact", "verb": "download", "positional": ["id"] },
        "x-auth": "session_or_api_key",
        "x-scopes": ["runs:read"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  },
  "security": [{ "apiKey": [] }, { "session": [] }]
}